fixes, for saving locations

+++ perl/openprint/photo_albums.pm
@@ -151,9 +151,11 @@ $log->debug("$id , $filename ");
 		$variable{'error'} .= $User->save({'asset_id'=>$param{'asset_id'}});
 	} elsif ( $param{'action'} eq 'delete' ) {
 		my $Asset = new openprint::Asset( $param{'asset_id'} );
-		foreach my $Photo ( openprint::Photo_in_Album->find( 'album_id'=>$$Album{'id'}, 'asset_id' => $Asset->id() ) ) {
+		foreach my $Photo ( $Album->Photos() ) {
+			next if $Photo->asset_id() != $Asset->id();
 			$variable{'error'} .= $Photo->delete();
 		} # end foreach Photo
+		$Album->Photos(undef); # Refresh cache
 
 		# Why am I deleting the asset?
 		#$variable{'error'} .= $Asset->delete();
